REL-208 refactor : issue refactoring
- class naming
