+++ src/main/java/com/momentum/releaser/domain/issue/mapper/IssueMapper.java
@@ -2,8 +2,8 @@ package com.momentum.releaser.domain.issue.mapper;
 
 import com.momentum.releaser.domain.issue.domain.Issue;
 import com.momentum.releaser.domain.issue.dto.IssueDataDto.ConnectedIssuesDataDto;
-import com.momentum.releaser.domain.issue.dto.IssueResDto.GetIssue;
-import com.momentum.releaser.domain.issue.dto.IssueResDto.OpinionInfoRes;
+import com.momentum.releaser.domain.issue.dto.IssueResponseDto.GetIssue;
+import com.momentum.releaser.domain.issue.dto.IssueResponseDto.OpinionInfoRes;
 import com.momentum.releaser.domain.project.dto.ProjectDataDto.GetMembers;
 import com.momentum.releaser.domain.project.mapper.ProjectMapper;
 import org.mapstruct.Mapper;
